Pull out 'matcher' interface
